% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6Methods.R
\name{Bottleneck_size_Calculation}
\alias{Bottleneck_size_Calculation}
\title{Calculate transmission bottleneck size}
\usage{
Bottleneck_size_Calculation(
  transmission_ob,
  method,
  plot = FALSE,
  show_table = FALSE,
  transmission_pairs = NULL,
  donor_depth_threshold = 10,
  recipient_depth_threshold = 10,
  error_filtering = 0,
  log = FALSE,
  variant_calling = 0.03,
  Nbmin = 1,
  Nbmax = 1000,
  NonSyn_or_Syn = "All"
)
}
\arguments{
\item{transmission_ob}{is a transmission object created by,`CreateTransmissionObject`.}

\item{method}{is characters containing `KL` method (Emmett et al., 2015), `Presence-Absence` method (Sacristán et al., 2003), `Binomial` method (Sobel Leonard et al., 2017), `Beta_binomial_Approximate` method (Sobel Leonard et al., 2017), `Beta_binomial_Exact` method (Sobel Leonard et al., 2017)and `Wright-Fisher` method(Poon et al., 2016)}

\item{plot}{is a logical value to determine plot the likelihood or not. Each plot would be stored in an individual folder named by transmission id.}

\item{show_table}{is a logical value to determine output a result table as `csv` format.}

\item{transmission_pairs}{is a dataframe which is the subset of the transmission pairs table during the object creation.}

\item{donor_depth_threshold}{is a integer to filter variants in donor with the lower sequencing coverage}

\item{recipient_depth_threshold}{is a integer to filter variants in recipient with the lower sequencing coverage}

\item{error_filtering}{is a numeric and filter. The proportion of the variants lower than error calling means the variant is difficult to distinguish from errors. After using that, the variants with values below this threshold would be zero.}

\item{log}{is a logical value to show the log or not. The log would be stored in an individual folder.}

\item{variant_calling}{is a threshold used for calling variant. The variants greater than this thrshould would be trusted.}

\item{Nbmin}{is a integer representing the minimum number in likelihood range.}

\item{Nbmax}{is a integer representing the maximum number in likelihood range.}

\item{NonSyn_or_Syn}{is a filter to keep the variant sites which meet the requirements. If user need the mutation of variant sites are synonymous mutation, then `NonSyn_or_Syn="Synonymous"`. If user need the mutation of variant sites are non-synonymous mutation, then `NonSyn_or_Syn="Non-Synonymous"`. The default is "All"}
}
\value{
Bottleneck size table containing the transmission pairs, bottleneck size, the lower confidence interval and higher confidence interval
}
\description{
Calculate transmission bottleneck size
}
\examples{
#The example can not be run directly
#Use different methods for estimation
#use KL method
KL_table = Bottleneck_size_calculation(transmission_object, method="KL")
#use presence-absence method
PA_table = Bottleneck_size_calculation(transmission_object, method="Presence-Absence")
#use Binomial method
Bi_table = Bottleneck_size_calculation(transmission_object, method="Binomial")
#use Beta-binomial approximate method
BB_App_table = Bottleneck_size_calculation(transmission_object, method="Beta_binomial_Approximate")
#use Beta-binomial approximate method
BB_Exact_table = Bottleneck_size_calculation(transmission_object, method="Beta_binomial_Exact")
#use wright-fisher method
BB_WF_table = Bottleneck_size_calculation(transmission_object, method="Wright-Fisher")
}
\references{
Emmett, K. J., Lee, A., Khiabanian, H., & Rabadan, R. (2015) 
High-resolution genomic surveillance of 2014 Ebolavirus using shared 
subclonal variants. \emph{PLOS Currents Outbreaks} \bold{7}, 
ecurrents.outbreaks.\cr 
Sacristán, S., Malpica, J. M., Fraile, A., & García-Arenal, F. (2003)
Estimation of population bottlenecks during systemic movement of
tobacco mosaic virus in tobacco plants. \emph{Journal of Virology}
\bold{77}(18), 9906–9911.\cr
Poon, L. L. M., Song, T., Rosenfeld, R., Lin, X., Rogers, M. B.,
Zhou, B., Sebra, R., Halpin, R., Guan, Y., Twaddle, A., DePasse, J.,
Stockwell, T., Wentworth, D., Holmes, E., Greenbaum, B., Peiris, J. S. M.,
Cowling, B. J., & Ghedin, E. (2016) Quantifying influenza virus
diversity and transmission in humans. \emph{Nature Genetics}
\bold{48}(2), 195--200.\cr
Sobel Leonard, A., Weissman, D. B., Greenbaum, B., Ghedin, E.,
& Koelle, K. (2017) Transmission bottleneck size estimation from
pathogen deep-sequencing data, with an application to human
influenza A virus. \emph{Journal of Virology} \bold{91}(14),
e00171-17.\cr
}
